improved stopping script

+++ dmon-stop.sh
@@ -10,13 +10,30 @@ elif [ $ARCH == "Darwin" ]; then
 fi
 
 echo "Stopping Kibana ..."
-kill -9  `cat $DIR/src/pid/kibana.pid`
+if [! -f $DIR/src/pid/kibana.pid]; then
+	echo "No Kibana instance to stop."
+else
+	kill -9  `cat $DIR/src/pid/kibana.pid`
+	echo "Kibana Stopped!"
+fi
+
 echo "Stopping Logstash Server ..."
-kill -9  `cat $DIR/src/pid/logstash.pid`
+if [! -f $DIR/src/pid/logstash.pid]; then
+	echo "No Logstash instance to stop."
+else
+	kill -9  `cat $DIR/src/pid/logstash.pid`
+	echo "Logstash Stopped!"
+fi
+
 echo "Stopping ElasticSearch ..."
-kill -9  `cat $DIR/src/pid/elasticsearch.pid`
+if [! -f $DIR/src/pid/elasticsearch.pid]; then
+	echo "No ElasticSearch instance to stop."
+else
+	kill -9  `cat $DIR/src/pid/elasticsearch.pid`
+	echo "ElasticSearch Stopped!"
+fi
 
 echo "Stopping D-Mon"
 killall -9 python #TODO need more elegant solution
 
-echo "D-Mon Stopped"
+echo "D-Mon Stopped!"
